Moved non view code to controller and handle Export only enabled when memory source is live feed.
